+++ src/main/java/com/github/demo/sms/SmsController.java
@@ -1,10 +1,10 @@
 package com.github.demo.sms;
 
-import com.github.jsoncat.annotation.Autowired;
-import com.github.jsoncat.annotation.PostMapping;
-import com.github.jsoncat.annotation.Qualifier;
-import com.github.jsoncat.annotation.RequestBody;
-import com.github.jsoncat.annotation.RestController;
+import com.github.jsoncat.annotation.ioc.Autowired;
+import com.github.jsoncat.annotation.springmvc.PostMapping;
+import com.github.jsoncat.annotation.ioc.Qualifier;
+import com.github.jsoncat.annotation.springmvc.RequestBody;
+import com.github.jsoncat.annotation.springmvc.RestController;
 
 /**
  * @author shuang.kou
