--- src/main/java/com/github/jsoncat/annotation/PathVariable.java
@@ -1,12 +1,0 @@
-package com.github.jsoncat.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface PathVariable {
-    String value();
-}
