+++ src/main/java/com/github/jsoncat/annotation/ioc/Autowired.java
@@ -1,4 +1,4 @@
-package com.github.jsoncat.annotation;
+package com.github.jsoncat.annotation.ioc;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,9 +6,9 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.TYPE)
+@Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface Component {
-    String name() default "";
+public @interface Autowired {
+
 }
