+++ src/main/java/com/github/jsoncat/annotation/ioc/Component.java
@@ -1,4 +1,4 @@
-package com.github.jsoncat.annotation;
+package com.github.jsoncat.annotation.ioc;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,13 +6,9 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/**
- * @author shuang.kou
- * @createTime 2020年09月24日 14:46:00
- **/
-@Target({ElementType.TYPE, ElementType.METHOD})
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface PostMapping {
-    String value() default "";
+public @interface Component {
+    String name() default "";
 }
