+++ src/main/java/com/github/jsoncat/annotation/ioc/Qualifier.java
@@ -1,4 +1,4 @@
-package com.github.jsoncat.annotation;
+package com.github.jsoncat.annotation.ioc;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
