+++ src/main/java/com/github/jsoncat/annotation/springmvc/PathVariable.java
@@ -1,17 +1,12 @@
-package com.github.jsoncat.annotation;
+package com.github.jsoncat.annotation.springmvc;
 
-
-import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-
-@Target({ElementType.TYPE})
+@Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Component
-public @interface RestController {
-    String value() default "";
+public @interface PathVariable {
+    String value();
 }
