+++ src/main/java/com/github/jsoncat/annotation/springmvc/PostMapping.java
@@ -1,0 +1,18 @@
+package com.github.jsoncat.annotation.springmvc;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * @author shuang.kou
+ * @createTime 2020年09月24日 14:46:00
+ **/
+@Target({ElementType.TYPE, ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface PostMapping {
+    String value() default "";
+}
