+++ src/main/java/com/github/jsoncat/annotation/springmvc/RequestBody.java
@@ -1,14 +1,11 @@
-package com.github.jsoncat.annotation;
+package com.github.jsoncat.annotation.springmvc;
 
-import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.FIELD)
+@Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface Autowired {
-
+public @interface RequestBody {
 }
