+++ src/main/java/com/github/jsoncat/annotation/springmvc/RequestParam.java
@@ -1,4 +1,4 @@
-package com.github.jsoncat.annotation;
+package com.github.jsoncat.annotation.springmvc;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,6 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface RequestBody {
+public @interface RequestParam {
+    String value();
 }
