+++ src/main/java/com/github/jsoncat/annotation/springmvc/RestController.java
@@ -1,0 +1,19 @@
+package com.github.jsoncat.annotation.springmvc;
+
+
+import com.github.jsoncat.annotation.ioc.Component;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Component
+public @interface RestController {
+    String value() default "";
+}
