+++ src/main/java/com/github/jsoncat/core/springmvc/factory/RouteMethodMapper.java
@@ -1,9 +1,12 @@
-package com.github.jsoncat.core.factory;
+package com.github.jsoncat.core.springmvc.factory;
 
-import com.github.jsoncat.annotation.GetMapping;
-import com.github.jsoncat.annotation.PostMapping;
-import com.github.jsoncat.annotation.RestController;
+import com.github.jsoncat.annotation.springmvc.GetMapping;
+import com.github.jsoncat.annotation.springmvc.PostMapping;
+import com.github.jsoncat.annotation.springmvc.RestController;
 import com.github.jsoncat.common.util.UrlUtil;
+import com.github.jsoncat.entity.MethodDetail;
+import com.github.jsoncat.factory.ClassFactory;
+import io.netty.handler.codec.http.HttpMethod;
 
 import java.lang.reflect.Method;
 import java.util.HashMap;
@@ -16,17 +19,17 @@ import java.util.Set;
  * @author shuang.kou
  * @createTime 2020年09月29日 13:27:00
  **/
-public class RouterFactory {
+public class RouteMethodMapper {
     // get request url -> target method.
     // eg: "^/user/[\u4e00-\u9fa5_a-zA-Z0-9]+/?$" -> UserController.get(java.lang.Integer)
-    public static final Map<String, Method> GET_REQUEST_MAPPINGS = new HashMap<>();
+    private static final Map<String, Method> GET_REQUEST_MAPPINGS = new HashMap<>();
     // post request url -> target method.
-    public static final Map<String, Method> POST_REQUEST_MAPPINGS = new HashMap<>();
+    private static final Map<String, Method> POST_REQUEST_MAPPINGS = new HashMap<>();
     // formatted get request url -> original url
     // eg : "^/user/[\u4e00-\u9fa5_a-zA-Z0-9]+/?$" -> /user/{id}
-    public static final Map<String, String> GET_URL_MAPPINGS = new HashMap<>();
+    private static final Map<String, String> GET_URL_MAPPINGS = new HashMap<>();
     // formatted post request url -> original url
-    public static final Map<String, String> POST_URL_MAPPINGS = new HashMap<>();
+    private static final Map<String, String> POST_URL_MAPPINGS = new HashMap<>();
 
     public static void loadRoutes() {
         Set<Class<?>> classes = ClassFactory.CLASSES.get(RestController.class);
@@ -57,6 +60,20 @@ public class RouterFactory {
                 }
             }
         }
+    }
+
+    public static MethodDetail getMethodDetail(String requestPath, HttpMethod httpMethod) {
+        MethodDetail methodDetail = new MethodDetail();
+        if (httpMethod == HttpMethod.GET) {
+            methodDetail.build(requestPath, GET_REQUEST_MAPPINGS, GET_URL_MAPPINGS);
+            return methodDetail;
+        }
+
+        if (httpMethod == HttpMethod.POST) {
+            methodDetail.build(requestPath, POST_REQUEST_MAPPINGS, POST_URL_MAPPINGS);
+            return methodDetail;
+        }
+        return null;
     }
 
 }
