+++ src/main/java/com/github/jsoncat/core/springmvc/handler/RequestHandler.java
@@ -1,4 +1,4 @@
-package com.github.jsoncat.core.handler;
+package com.github.jsoncat.core.springmvc.handler;
 
 import io.netty.handler.codec.http.FullHttpRequest;
