+++ src/main/java/com/github/jsoncat/core/springmvc/resolver/ParameterResolver.java
@@ -1,6 +1,6 @@
-package com.github.jsoncat.core.resolver;
+package com.github.jsoncat.core.springmvc.resolver;
 
-import com.github.jsoncat.core.entity.MethodDetail;
+import com.github.jsoncat.entity.MethodDetail;
 
 import java.lang.reflect.Parameter;
