+++ src/main/java/com/github/jsoncat/entity/MethodDetail.java
@@ -1,0 +1,41 @@
+package com.github.jsoncat.entity;
+
+import com.github.jsoncat.common.util.UrlUtil;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.lang.reflect.Method;
+import java.util.Map;
+import java.util.regex.Pattern;
+
+/**
+ * @author shuang.kou
+ * @createTime 2020年09月28日 09:42:00
+ **/
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MethodDetail {
+    // target method
+    private Method method;
+    // url parameter mapping
+    private Map<String, String> urlParameterMappings;
+    // url query parameter mapping
+    private Map<String, String> queryParameterMappings;
+    // json type http post request data
+    private String json;
+
+    public void build(String requestPath, Map<String, Method> requestMappings, Map<String, String> urlMappings) {
+        requestMappings.forEach((key, value) -> {
+            Pattern pattern = Pattern.compile(key);
+            boolean b = pattern.matcher(requestPath).find();
+            if (b) {
+                this.setMethod(value);
+                String url = urlMappings.get(key);
+                Map<String, String> urlParameterMappings = UrlUtil.getUrlParameterMappings(requestPath, url);
+                this.setUrlParameterMappings(urlParameterMappings);
+            }
+        });
+    }
+}
