+++ src/test/java/com/github/jsoncat/core/scanner/AnnotatedClassScannerTest.java
@@ -1,6 +1,7 @@
 package com.github.jsoncat.core.scanner;
 
-import com.github.jsoncat.annotation.RestController;
+import com.github.jsoncat.annotation.springmvc.RestController;
+import com.github.jsoncat.common.util.ReflectionUtil;
 import org.junit.jupiter.api.Test;
 
 import java.util.Set;
@@ -12,7 +13,7 @@ class AnnotatedClassScannerTest {
 
     @Test
     void should_scan_the_annotated_class() {
-        Set<Class<?>> annotatedClasses = AnnotatedClassScanner.scan("com.github.demo", RestController.class);
+        Set<Class<?>> annotatedClasses = ReflectionUtil.scanAnnotatedClass("com.github.demo", RestController.class);
         assertEquals(2, annotatedClasses.size());
     }
